Refactor to use AnyShapeStyle extension
