+++ Sources/LiveViewNative/Modifiers/ForegroundStyleModifier.swift
@@ -8,27 +8,18 @@
 import SwiftUI
 
 struct ForegroundStyleModifier: ViewModifier, Decodable {
-    private let primary: WrappedShapeStyle?
+    private let primary: AnyShapeStyle?
 
     init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
 
-        self.primary = try container.decodeIfPresent(WrappedShapeStyle.self, forKey: .primary)
+        self.primary = try container.decodeIfPresent(AnyShapeStyle.self, forKey: .primary)
     }
 
     func body(content: Content) -> some View {
         let primaryStyle = self.primary!
 
-        switch primaryStyle.concreteStyle {
-        case .color:
-            content.foregroundStyle(primaryStyle.color!)
-            
-        case .linearGradient:
-            content.foregroundStyle(primaryStyle.linearGradient!)
-            
-        default:
-            content
-        }
+        content.foregroundStyle(primaryStyle)
     }
     
     enum CodingKeys: String, CodingKey {
