+++ Sources/LiveViewNative/Views/Drawing and Graphics/Gradient/Stop.swift
@@ -10,7 +10,7 @@ import SwiftUI
 extension Gradient.Stop: Decodable {
     public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
-        let color: Color = try container.decode(Color.self, forKey: .color)
+        let color: SwiftUI.Color = try container.decode(SwiftUI.Color.self, forKey: .color)
         let location: CGFloat = try container.decode(CGFloat.self, forKey: .location)
     
         self = Gradient.Stop(color: color, location: location)
