+++ Sources/LiveViewNative/Views/Drawing and Graphics/LinearGradient.swift
@@ -4,7 +4,6 @@
 //
 //  Created by May Matyi on 2/17/23.
 //
-
 import SwiftUI
 
 extension LinearGradient: Decodable {
@@ -13,10 +12,10 @@ extension LinearGradient: Decodable {
         let gradient: Gradient = try container.decode(Gradient.self, forKey: .gradient)
         let startPoint: UnitPoint = try container.decode(UnitPoint.self, forKey: .startPoint)
         let endPoint: UnitPoint = try container.decode(UnitPoint.self, forKey: .endPoint)
-    
+
         self = LinearGradient(gradient: gradient, startPoint: startPoint, endPoint: endPoint)
     }
-    
+
     enum CodingKeys: String, CodingKey {
         case gradient
         case startPoint = "start_point"
